split tcp_proxy test into config test and data test
