+++ test/common/tcp_proxy/BUILD
@@ -1,6 +1,7 @@
 load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_test",
+    "envoy_cc_test_library",
     "envoy_package",
 )
 
@@ -8,9 +9,11 @@ licenses(["notice"])  # Apache 2
 
 envoy_package()
 
-envoy_cc_test(
-    name = "tcp_proxy_test",
-    srcs = ["tcp_proxy_test.cc"],
+envoy_cc_test_library(
+    name = "tcp_proxy_test_base",
+    srcs = [
+        "tcp_proxy_test_base.h",
+    ],
     deps = [
         "//source/common/buffer:buffer_lib",
         "//source/common/event:dispatcher_lib",
@@ -33,8 +36,34 @@ envoy_cc_test(
         "//test/mocks/server:instance_mocks",
         "//test/mocks/ssl:ssl_mocks",
         "//test/mocks/stream_info:stream_info_mocks",
+        "//test/mocks/upstream:cluster_manager_mocks",
         "//test/mocks/upstream:host_mocks",
         "//test/test_common:test_runtime_lib",
+        "@envoy_api//envoy/config/accesslog/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/access_loggers/file/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/filters/network/tcp_proxy/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/upstreams/http/generic/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/upstreams/tcp/generic/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_test(
+    name = "config_test",
+    srcs = [
+        "config_test.cc",
+    ],
+    deps = [
+        ":tcp_proxy_test_base",
+    ],
+)
+
+envoy_cc_test(
+    name = "tcp_proxy_test",
+    srcs = [
+        "tcp_proxy_test.cc",
+    ],
+    deps = [
+        ":tcp_proxy_test_base",
         "@envoy_api//envoy/config/accesslog/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/access_loggers/file/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/filters/network/tcp_proxy/v3:pkg_cc_proto",
@@ -50,5 +79,6 @@ envoy_cc_test(
         "//source/common/tcp_proxy",
         "//test/mocks/http:http_mocks",
         "//test/mocks/tcp:tcp_mocks",
+        "//test/mocks/upstream:cluster_manager_mocks",
     ],
 )
